+++ models/oneinch/ethereum/oneinch_v3_ethereum_trades.sql
@@ -150,7 +150,7 @@ SELECT
     ,src.tx_hash
     ,tx.from AS tx_from
     ,tx.to AS tx_to
-    ,CAST(src.trace_address as array<long>) as trace_address
+    ,CAST(src.trace_address as array<numeric>) as trace_address
     ,src.evt_index
 FROM oneinch as src
 INNER JOIN {{ source('ethereum', 'transactions') }} as tx
